chat-app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-SocketIO==5.3.6
groq==0.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.43.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
zipp==3.20.1
